move to attention dir
